# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  ginutil.Result:
    properties:
      code:
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
...
  v1.PruneUpdateRKEConfigReq:
    properties:
      encodedRKEConfig:
        type: string
      nodes:
        items:
          $ref: '#/definitions/v1alpha1.ConfigNode'
        type: array
    type: object
  v1.PruneUpdateRKEConfigResp:
    properties:
      encodeRKEConfig:
        type: string
      nodes:
        items:
          $ref: '#/definitions/v1alpha1.ConfigNode'
        type: array
    type: object
...
  v1.WutongComponent:
    properties:
      app:
        type: string
      pods:
        items:
          $ref: '#/definitions/v1.Pod'
        type: array
    type: object
...
  v1.UpdateKubernetesReq:
    properties:
      clusterID:
        type: string
      encodedRKEConfig:
        type: string
      etcdNodeNum:
        type: integer
      instanceType:
        type: string
      masterNodeNum:
        type: integer
      provider:
        type: string
      workerNum:
        type: integer
      workerResourceType:
        type: string
    type: object
  v1.UpdateKubernetesTask:
    properties:
      clusterID:
        type: string
      eid:
        type: string
      nodeNumber:
        type: integer
      providerName:
        type: string
      status:
        type: string
      taskID:
        type: string
    type: object
...
  v1alpha1.ConfigNode:
    properties:
      dockerSocketPath:
        type: string
      internalIP:
        type: string
      ip:
        type: string
      roles:
        items:
          type: string
        type: array
      sshPort:
        type: integer
      sshUser:
        type: string
    type: object
...
info:
  contact:
    name: Wutong
    url: https://wutong.com
  description: Cloud Adaptor
  title: Cloud Adaptor API
  version: "1.0"
paths:
...
  /api/v1/enterprises/:eid/kclusters/prune-update-rkeconfig:
    post:
      consumes:
      - application/json
      operationId: pruneUpdateRKEConfig
      parameters:
      - description: the enterprise id
        in: path
        name: eid
        required: true
        type: string
      - description: .
        in: body
        name: pruneUpdateRKEConfigReq
        required: true
        schema:
          $ref: '#/definitions/v1.PruneUpdateRKEConfigReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.PruneUpdateRKEConfigResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ginutil.Result'
      summary: update rke config purely
      tags:
      - cluster
  /api/v1/enterprises/:eid/update-cluster:
    post:
      consumes:
      - application/json
      operationId: updateKubernetesCluster
      parameters:
      - description: the enterprise id
        in: path
        name: eid
        required: true
        type: string
      - description: .
        in: body
        name: updateKubernetesReq
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateKubernetesReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.UpdateKubernetesTask'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ginutil.Result'
      summary: updates kubernetes cluster.
      tags:
      - cluster
  /api/v1/enterprises/:eid/update-cluster/:clusterID:
    get:
      consumes:
      - application/json
      operationId: getUpdateKubernetesTask
      parameters:
      - description: the enterprise id
        in: path
        name: eid
        required: true
        type: string
      - description: the cluster id
        in: path
        name: clusterID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.UpdateKubernetesTask'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ginutil.Result'
      summary: returns the information of the cluster.
      tags:
      - cluster
  /api/v1/enterprises/{eid}/kclusters/{clusterID}/wutong-components:
    get:
      consumes:
      - application/json
      operationId: listWutongComponents
      parameters:
      - description: the enterprise id
        in: path
        name: eid
        required: true
        type: string
      - description: the identify of cluster
        in: path
        name: clusterID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.WutongComponent'
            type: array
      summary: returns a list of wutong components.
      tags:
      - cluster
  /api/v1/enterprises/{eid}/kclusters/{clusterID}/wutong-components/{podName}/events:
    get:
      consumes:
      - application/json
      operationId: listPodEvents
      parameters:
      - description: the enterprise id
        in: path
        name: eid
        required: true
        type: string
      - description: the identify of cluster
        in: path
        name: clusterID
        required: true
        type: string
      - description: the name of pod
        in: path
        name: podName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.WutongComponentEvent'
            type: array
      summary: returns a list of wutong component pod events.
      tags:
      - cluster
schemes:
- http,
- https
swagger: "2.0"